Fix : getFileSize returns wrong size for >= 8Gb files... (So they are hashed before scrapping which is eternal )

+++ es-core/src/utils/FileSystemUtil.h
@@ -60,7 +60,7 @@ namespace Utils
 
 		fileList	getDirectoryFiles(const std::string& _path);
 		std::string combine(const std::string& _path, const std::string& filename);
-		size_t		getFileSize(const std::string& _path);
+		unsigned long long	getFileSize(const std::string& _path);
 
 		Utils::Time::DateTime getFileCreationDate(const std::string& _path);
 		Utils::Time::DateTime getFileModificationDate(const std::string& _path);
